requirements.txt
algoliasearch==2.6.2
algoliasearch-django==2.0.0
aniso8601==9.0.1
anyio==3.6.2
asgiref==3.6.0
attrs==22.2.0
autopep8==2.0.0
bcrypt==4.0.1
blessed==1.19.1
boto3==1.26.41
botocore==1.29.41
cached-property==1.5.2
cachetools==5.2.0
cement==2.8.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.3
commonmark==0.9.1
configparser==5.3.0
cryptography==38.0.4
Django==4.1.3
django-choices-field==2.1
django-cleanup==6.0.0
django-cors-headers==3.13.0
django-ebhealthcheck==2.0.2
django-extensions==3.2.1
django-filter==22.1
django-storages==1.13.2
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
dnspython==2.2.1
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
Faker==15.3.4
future==0.18.3
gprof2dot==2022.7.29
graphene==3.1.1
graphene-django==3.0.0
graphene-file-upload==1.3.0
graphql-core==3.2.3
graphql-relay==3.2.0
guardian==0.2.1
gunicorn==20.1.0
h11==0.14.0
idna==3.4
iso8601==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
jsonschema==3.2.0
libcst==0.4.9
Markdown==3.4.1
MarkupSafe==2.1.1
mypy-extensions==0.4.3
paramiko==2.12.0
pathspec==0.9.0
Pillow==9.3.0
promise==2.3
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pycparser==2.21
Pygments==2.14.0
PyNaCl==1.5.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2022.6
PyYAML==5.4.1
requests==2.26.0
rich==13.0.1
s3transfer==0.6.0
semantic-version==2.8.5
singledispatch==3.7.0
six==1.14.0
sniffio==1.3.0
sqlparse==0.2.4
starlette==0.23.1
strawberry-graphql==0.209.2
strawberry-graphql-django==0.17.4
termcolor==1.1.0
text-unidecode==1.3
texttable==1.6.7
tomli==2.0.1
typer==0.9.0
typing-inspect==0.8.0
typing_extensions==4.8.0
urllib3==1.26.13
uvicorn==0.20.0
wcwidth==0.1.9
websocket-client==0.59.0
